Split code by domain not by nature
