+++ app/assets/javascripts/application/cookies_warning.component.js
