+++ app/assets/javascripts/application/email_notifications_reminder.component.js
