+++ app/assets/javascripts/application/loading.component.js
