+++ app/assets/javascripts/application/pagination.component.js
