+++ app/assets/javascripts/categories/subcategory_picker.component.js
