+++ app/assets/javascripts/components.js
@@ -1,39 +1,35 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React                      from 'react';
+import ReactDOM                   from 'react-dom';
+import Loading                    from './application/loading.component';
+import CookiesWarning             from './application/cookies_warning.component';
+import EmailNotificationsReminder from './application/email_notifications_reminder.component';
+import StaticMap                  from './location/static_map.component';
+import AutocompleteInputAddress   from './location/autocomplete_input_address.component';
+import FilterService              from './filters/filter.service';
+import CategoryPicker             from './categories/category_picker.component';
+import ProposalsCarousel          from './proposals/proposals_carousel.component';
+import ProposalFilterTabs         from './proposals/proposal_filter_tabs.component';
+import ProposalFilters            from './proposals/proposal_filters.component';
+import Votes                      from './proposals/votes.component';
+import MeetingsDirectory          from './meetings/meetings_directory.component';
+import MeetingsCarousel           from './meetings/meetings_carousel.component';
+import MeetingsMap                from './meetings/meetings_map.component';
+import MeetingProposalsSelector   from './meetings/meeting_proposals_selector.component';
 
-import FilterService from './services/filter_service';
-
-import MeetingsDirectory from './components/meetings_directory';
-import StaticMap from './components/static_map';
-import ProposalsCarousel from './components/proposals_carousel';
-import MeetingsCarousel from './components/meetings_carousel';
-import MeetingsMap from './components/meetings_map';
-import ProposalFilterTabs from './components/proposal_filter_tabs';
-import ProposalFilters from './components/proposal_filters';
-import Loading from './components/loading';
-import AutocompleteInputAddress from './components/autocomplete_input_address';
-import MeetingProposalsSelector from './components/meeting_proposals_selector';
-import CategoryPicker from './components/category_picker';
-import CookiesWarning from './components/cookies_warning';
-import EmailNotificationsReminder from './components/email_notifications_reminder';
-import Votes from './components/votes';
-
-window.React = React;
-window.ReactDOM = ReactDOM;
-
-window.FilterServiceInstance = new FilterService(); 
-
-window.MeetingsDirectory = MeetingsDirectory;
-window.StaticMap = StaticMap;
-window.ProposalsCarousel = ProposalsCarousel;
-window.MeetingsCarousel = MeetingsCarousel;
-window.MeetingsMap = MeetingsMap;
-window.ProposalFilterTabs = ProposalFilterTabs;
-window.ProposalFilters = ProposalFilters;
-window.Loading = Loading;
-window.AutocompleteInputAddress = AutocompleteInputAddress;
-window.MeetingProposalsSelector = MeetingProposalsSelector;
-window.CategoryPicker = CategoryPicker;
-window.CookiesWarning = CookiesWarning;
+window.React                      = React;
+window.ReactDOM                   = ReactDOM;
+window.Loading                    = Loading;
+window.CookiesWarning             = CookiesWarning;
 window.EmailNotificationsReminder = EmailNotificationsReminder;
-window.Votes = Votes;
+window.StaticMap                  = StaticMap;
+window.AutocompleteInputAddress   = AutocompleteInputAddress;
+window.FilterServiceInstance      = new FilterService();
+window.CategoryPicker             = CategoryPicker;
+window.ProposalsCarousel          = ProposalsCarousel;
+window.ProposalFilterTabs         = ProposalFilterTabs;
+window.ProposalFilters            = ProposalFilters;
+window.Votes                      = Votes;
+window.MeetingsDirectory          = MeetingsDirectory;
+window.MeetingsCarousel           = MeetingsCarousel;
+window.MeetingsMap                = MeetingsMap;
+window.MeetingProposalsSelector   = MeetingProposalsSelector;
