+++ app/assets/javascripts/filters/filter.service.js
