+++ app/assets/javascripts/filters/filter_option.component.js
