+++ app/assets/javascripts/filters/filter_option_group.component.js
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 
-import FilterOption from './filter_option';
+import FilterOption  from './filter_option.component';
 
 export default class FilterOptionGroup extends Component {
   constructor(props) {
