+++ app/assets/javascripts/filters/search_filter.component.js
