+++ app/assets/javascripts/filters/tag_cloud_filter.component.js
