+++ app/assets/javascripts/location/static_map.component.js
