+++ app/assets/javascripts/meetings/meeting_proposals_table.component.js
