+++ app/assets/javascripts/proposals/votes.component.js
